+++ Raytracer.jl
@@ -26,6 +26,7 @@ import FileIO: @format_str, query
 using Raytracing
 
 FILE_NAME = split(PROGRAM_FILE, "/")[end]
+RENDERERS = ["onoff", "flat", "pathtracing"]
 
 function parse_commandline_error_handler(settings::ArgParseSettings, err, err_code::Int = 1)
 	help_string = 
@@ -57,23 +58,24 @@ end
 
 function ArgParse_command_line(arguments)
 	s = ArgParseSettings()
+
 	s.description = "Raytracer for the generation of photorealistic images in Julia."
 	s.exc_handler = parse_commandline_error_handler
 	s.version = @project_version
-	#s.version = "0.1.0"
+
 	@add_arg_table! s begin
 		"demo"
 			action = :command
-			help = "create a standard image that checks the correct behaviour of the program."
+			help = "create a standard image that checks the correct behaviour of the program"
 		"demo-animation"
 			action = :command
-			help = "create an animation of a 360 degree rotation of demo image "
+			help = "create an animation of a 360 degree rotation of the demo image"
 		"tonemapping"
 			action = :command
-			help = "apply tone mapping to a pfm image and save it to file"
+			help = "apply tone mapping to a pfm image and save it as a ldr file"
 	end
 
-	s["tonemapping"].description = "Apply tone mapping to a pfm image and save it to file."
+	s["tonemapping"].description = "Apply tone mapping to a pfm image and save it as a ldr file"
 	add_arg_group!(s["tonemapping"], "tonemapping settings");
 	@add_arg_table! s["tonemapping"] begin
 		"--alpha", "-a"
@@ -85,7 +87,7 @@ function ArgParse_command_line(arguments)
 			arg_type = Float64
 			default = 1.27
 	end
-	add_arg_group!(s["tonemapping"], "filenames");
+	add_arg_group!(s["tonemapping"], "tonemapping filenames");
 	@add_arg_table! s["tonemapping"] begin
 		"pfm_infile"
 			help = "path to input file, it must be a PFM file"
@@ -96,10 +98,11 @@ function ArgParse_command_line(arguments)
 			required = true
 	end
 
-	s["demo"].description = "creates a demo.png and a demo.pfm files of 10 white spheres on a dark backgroud."*
+	s["demo"].description = "Creates a demo.png and a demo.pfm files of 10 white spheres on a dark backgroud. "*
 					"8 spheres are placed at the 8 vertexes of a cube, 1 in the middle "*
-					"of the left face and the last in the lower one."*
+					"of the left face and the last in the lower one. "*
 					"Execute it in order to check the correct placement of the spheres."
+	#=
 	add_arg_group!(s["demo"], "demo mutually exclusive flags for the viewpoint tipe", exclusive=true)
     	@add_arg_table! s["demo"] begin
     		"--orthogonal", "--ort"
@@ -109,8 +112,41 @@ function ArgParse_command_line(arguments)
 			help = "flag for the Perspective projection view"
 			action = :store_true
     	end
-	add_arg_group!(s["demo"], "demo angle viewpoint direction and image dimensions");
+	=#
+
+    	
+	add_arg_group!(s["demo"], "demo options");
 	@add_arg_table! s["demo"] begin
+		"--camera-type", "-t"
+			help = "option for the camera type:\n"*
+	    				"ort -> Orthogonal camera, per -> Perspective camera"
+          	arg_type = String
+			default = "per"
+			range_tester = input -> (input ∈ ["ort", "per"])
+    		"--algorithm", "-r"
+			help = "option for the renderer algorithm"
+          	arg_type = String
+			default = "onoff"
+			range_tester = input -> (input ∈ RENDERERS)
+		"--world-type", "-l"
+			help = "flag for the world to be rendered"
+          	arg_type = String
+			default = "A"
+			range_tester = input -> (input ∈ ["A", "B"])
+    		"--init-state"
+    			arg_type=Int64
+    			help="Initial seed for the random number generator (positive number)."
+    			default=45
+    		"--init-seq"
+    			arg_type=Int64
+    			help="Identifier of the sequence produced by the "*
+			    "random number generator (positive number)."
+    			default=54
+		"--camera-position", "-p"
+          	help = "camera position in the scene as 'X,Y,Z'"
+          	arg_type = String
+          	default = "-1,0,0"
+          	range_tester = input -> (length(split(input, ",")) == 3)
 		"--alpha", "-a"
 			help = "angle of view, in degrees"
 			arg_type = Float64
@@ -125,7 +161,6 @@ function ArgParse_command_line(arguments)
 			arg_type = Int64
 			default = 480
 			range_tester = iseven
-
 	end
 	add_arg_group!(s["demo"], "demo optional filenames");
 	@add_arg_table! s["demo"] begin
@@ -145,17 +180,19 @@ function ArgParse_command_line(arguments)
 
 	s["demo-animation"].description = "creates an animation of a 360 degree rotation around"*
 								"vertical axis of the demo image."
-	add_arg_group!(s["demo-animation"], "demo-animation mutually exclusive flags for the viewpoint tipe", exclusive=true)
-    @add_arg_table! s["demo-animation"] begin
-    	"--orthogonal", "--ort"
-          	action = :store_true
-          	help = "flag for the Orthogonal projection view"
-		"--perspective", "--per"
-			help = "flag for the Perspective projection view"
-			action = :store_true
-    	end
 	add_arg_group!(s["demo-animation"], "demo-animation settings");
 	@add_arg_table! s["demo-animation"] begin
+		"--camera-type", "-t"
+			help = "flag for the camera type:\n"*
+	    				"ort -> Orthogonal camera, per -> Perspective camera"
+          	arg_type = String
+			default = "per"
+			range_tester = input -> (input ∈ ["ort", "per"])
+    		"--algorithm", "-r"
+			help = "flag for the renderer algorithm"
+          	arg_type = String
+			default = "onoff"
+			range_tester = input -> (input ∈ RENDERERS)
 		"--width", "-o"
 			help = "pixel number on the width of the resulting demo image animation."
 			arg_type = Int64