+++ src/Shapes.jl
@@ -4,17 +4,7 @@
 #
 # Copyright © 2021 Matteo Foglieni and Riccardo Gervasoni
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-# documentation files (the “Software”), to deal in the Software without restriction, including without limitation the
-# rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
-# the Software. THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
-# LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
-# SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
-# CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-# IN THE SOFTWARE.
+
 
 ##########################################################################################92
 
@@ -128,7 +118,8 @@ function ray_intersection(sphere::Sphere, ray::Ray)
         sphere.T * sphere_normal(hit_point, inv_ray.dir),
         sphere_point_to_uv(hit_point),
         first_hit_t,
-        ray
+        ray, 
+        sphere
     )
 end
 
@@ -156,7 +147,8 @@ function ray_intersection(plane::Plane, ray::Ray)
         plane.T * plane_normal(hit_point, inv_ray.dir),
         plane_point_to_uv(hit_point),
         hit_t,
-        ray
+        ray,
+        plane
     )
 end
 
@@ -164,13 +156,13 @@ end
 ##########################################################################################92
 
 """
-    add_shape(W::World, S::Shape)
+    add_shape!(W::World, S::Shape)
 
 Append a new shape to this world
 
 See also: [`Shape`](@ref), [`World`](@ref)
 """
-function add_shape(W::World, S::Shape)
+function add_shape!(W::World, S::Shape)
     push!(W.shapes, S)
     return nothing
 end