+++ src/Structs.jl
@@ -4,22 +4,16 @@
 #
 # Copyright © 2021 Matteo Foglieni and Riccardo Gervasoni
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-# documentation files (the “Software”), to deal in the Software without restriction, including without limitation the
-# rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
-# the Software. THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
-# LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
-# SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
-# CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-# IN THE SOFTWARE.
+
 
 ##########################################################################################92
 
 BLACK = RGB{Float32}(0.0, 0.0, 0.0)
 WHITE = RGB{Float32}(1.0, 1.0, 1.0)
+
+function to_RGB(r::Int64, g::Int64, b::Int64)
+    return RGB{Float32}(r/255., g/255., b/255.)
+end
 
 ##########################################################################################92
 
@@ -126,6 +120,10 @@ struct Normal
         m = √(x^2+y^2+z^2)
         new(x/m, y/m, z/m)
     end
+    function Normal(v::Vec)
+        m = √(v.x^2+v.y^2+v.z^2)
+        new(v.x/m, v.y/m, v.z/m)
+    end
     function Normal(v::Vector{Float64})
         @assert length(v) == 3
         m = √(v[1]^2+v[2]^2+v[3]^2)
@@ -136,6 +134,7 @@ struct Normal
         new(v[1]/m, v[2]/m, v[3]/m)
     end
 end
+Vec(N::Normal) = Vec(N.x, N.y, N.z)
 
 ##########################################################################################92
 
@@ -229,27 +228,7 @@ end
 
 abstract type Shape end
 
-##########################################################################################92
 
-"""
-A 3D unit sphere centered on the origin of the axes
-# Arguments
-- `T`: potentially [`Transformation`](@ref) associated to the sphere
-"""
-struct Sphere <: Shape
-    T::Transformation
-    Sphere(T=Transformation()) = new(T)
-end
-
-"""
-A 3D unit plane, i.e. the x-y plane (set of 3D points with z=0)
-# Arguments
-- `T`: potentially [`Transformation`](@ref) associated to the plane
-"""
-struct Plane <: Shape
-    T::Transformation
-    Plane(T=Transformation()) = new(T)
-end
 
 ##########################################################################################92
 
@@ -280,7 +259,8 @@ struct HitRecord
     surface_point::Vec2d
     t::Float64
     ray::Ray
-    HitRecord(w,n,s,t,r) =  new(w,n,s,t,r)
+    shape::Union{Shape, Nothing}
+    HitRecord(w,n,s,t,r, shp=nothing) =  new(w,n,s,t,r, shp)
     #=
     function HitRecord(w,n,s,t,r) 
         norm = normalize(n)
@@ -293,7 +273,7 @@ end
 
 """
 A struct holding a list of shapes, which make a «world»
-You can add shapes to a world using [`add_shape`](@ref)([`World`](@ref), [`Shape`](@ref)).
+You can add shapes to a world using [`add_shape!`](@ref)([`World`](@ref), [`Shape`](@ref)).
 Typically, you call [`ray_intersection`](@ref)([`World`](@ref), [`Ray`](@ref))
 to check whether a light ray intersects any of the shapes in the world.
 """
@@ -301,4 +281,196 @@ struct World
     shapes::Array{Shape}
     World(s::Shape) = new(s)
     World() = new( Array{Shape,1}() )
+end
+
+##########################################################################################92
+
+"""
+A «pigment»
+
+This abstract class represents a pigment, i.e., a function that associates a color with
+each point on a parametric surface (u,v). Call the method :meth:`.Pigment.get_color` to
+retrieve the color of the surface given a :class:`.Vec2d` object.
+"""
+abstract type Pigment end
+
+"""
+A uniform pigment
+This is the most boring pigment: a uniform hue over the whole surface.
+"""
+struct UniformPigment <: Pigment
+    color::RGB{Float32}
+    UniformPigment(c = BLACK) = new(c)
+end
+
+"""
+A checkered pigment
+The number of rows/columns in the checkered pattern is tunable, but you cannot have a different number of
+repetitions along the u/v directions.
+"""
+struct CheckeredPigment <: Pigment
+    color1::RGB{Float32}
+    color2::RGB{Float32}
+    num_steps::Int64
+    CheckeredPigment(c1 = WHITE, c2 = BLACK, n = 2) = new(c1, c2, n)
+end
+
+"""
+A textured pigment
+The texture is given through a PFM image.
+"""
+struct ImagePigment <: Pigment
+    image::HDRimage
+    ImagePigment(img = HDRimage(3, 2, fill(BLACK, (6,)))) = new(img)
+end
+
+##########################################################################################92
+
+"""
+An abstract class representing a Bidirectional Reflectance Distribution Function
+"""
+abstract type BRDF end
+
+"""
+A class representing an ideal diffuse BRDF (also called «Lambertian»)
+"""    
+struct DiffuseBRDF <: BRDF
+    pigment::Pigment
+    reflectance::Float64
+    DiffuseBRDF(pig = UniformPigment(WHITE), r=1.0) = new(pig, r)
+end
+
+"""
+A class representing an ideal mirror BRDF
+"""
+struct SpecularBRDF <: BRDF
+    pigment::Pigment
+    theresold_angle_rad::Float64
+    SpecularBRDF(p=UniformPigment(WHITE), thAngle=π/180.) = new(p, thAngle)
+end
+
+"""
+A material
+"""
+struct Material
+    brdf::BRDF
+    emitted_radiance::Pigment
+    Material(brdf = DiffuseBRDF(), er = UniformPigment()) = new(brdf, er)
+end
+
+##########################################################################################92
+
+"""
+A 3D unit sphere centered on the origin of the axes
+# Arguments
+- `T`: potentially [`Transformation`](@ref) associated to the sphere
+- `Material`: potentially [`Material`](@ref) associated to the sphere
+"""
+struct Sphere <: Shape
+    T::Transformation
+    Material::Material
+    
+    Sphere(T::Transformation, M::Material) = new(T,M)
+    Sphere(M::Material, T::Transformation) = new(T,M)
+    Sphere(T::Transformation) = new(T, Material())
+    Sphere(M::Material) = new(Transformation(), M)
+    Sphere() = new(Transformation(), Material())
+    
+    #Sphere(T=Transformation(), M=Material()) = new(T,M)
+end
+
+"""
+A 3D unit plane, i.e. the x-y plane (set of 3D points with z=0)
+# Arguments
+- `T`: potentially [`Transformation`](@ref) associated to the plane
+- `Material`: potentially [`Material`](@ref) associated to the plane
+"""
+struct Plane <: Shape
+    T::Transformation
+    Material::Material
+    
+    Plane(T::Transformation, M::Material) = new(T,M)
+    Plane(M::Material, T::Transformation) = new(T,M)
+    Plane(T::Transformation) = new(T, Material())
+    Plane(M::Material) = new(Transformation(), M)
+    Plane() = new(Transformation(), Material())
+    #Plane(T=Transformation(), M=Material()) = new(T,M)
+end
+
+##########################################################################################92
+
+"""
+A class implementing a solver of the rendering equation.
+This is an abstract class; you should use a derived concrete class.
+"""
+abstract type Renderer <: Function end
+
+"""
+A on/off renderer
+This renderer is mostly useful for debugging purposes, 
+as it is really fast, but it produces boring images.
+"""
+struct OnOffRenderer <: Renderer
+    world::World
+    background_color::RGB{Float32}
+    color::RGB{Float32}
+    OnOffRenderer(w = World(), bc = BLACK, c = WHITE) = new(w, bc, c)
+end
+
+"""
+A «flat» renderer
+This renderer estimates the solution of the rendering equation by neglecting any contribution of the light.
+It just uses the pigment of each surface to determine how to compute the final radiance.
+"""
+struct FlatRenderer <: Renderer
+    world::World
+    background_color::RGB{Float32}
+    FlatRenderer(w = World(), bc = BLACK) = new(w, bc)
+end
+
+"""
+    PathTracer(
+            world::World, 
+            background_color::RGB{Float32} = BLACK,
+            pcg::PCG = PCG(),
+            N::Int64 = 10,
+            max_depth::Int64 = 2,
+            russian_roulette_limit::Int64 = 3
+        )
+
+A simple path-tracing renderer.
+
+The algorithm implemented here allows the caller to tune number 
+of rays thrown at each iteration, as well as the maximum depth. 
+It implements Russian roulette, so in principle it will take a 
+finite time to complete the calculation even if you set 
+max_depth to `Inf`.
+
+## Arguments
+
+- `world::World` : the world to be rendered
+
+- `background_color::RGB{Float32}` : default background color 
+  if the Ray doesn-t hit anything
+
+- `pcg::PCG` :  PCG random number generator for evaluating integrals
+
+- `num_of_rays::Int64` : number of `Ray`s generated for each integral evaluation
+
+- `max_depth::Int64` : maximal number recursive integrations
+
+- `russian_roulette_limit::Int64`: depth at whitch the Russian 
+  Roulette algorithm begins
+
+See also: [`Ray`](@ref), [`World`](@ref), [`PCG`](@ref)
+"""
+struct PathTracer <: Renderer
+    world::World
+    background_color::RGB{Float32}
+    pcg::PCG
+    num_of_rays::Int64
+    max_depth::Int64
+    russian_roulette_limit::Int64
+    PathTracer(w, bc=BLACK, pcg=PCG(), n=10, md=2, RRlim=3) = 
+        new(w, bc, pcg, n, md, RRlim)
 end